requirements.txt
pandas
google-play-scraper
google-generativeai
tqdm
scikit-learn
python-dotenv
